demrepofdave/al5int/resize_windows: Better screen ratio.  Tidy-ups.  Old windows code archived to old_src directory for now
