+++ .vscode/settings.json
@@ -20,6 +20,7 @@
         "menu_internal.h": "c",
         "zlib.h": "c",
         "stdio.h": "c",
-        "logger.h": "c"
+        "logger.h": "c",
+        "ula.h": "c"
     }
 }
