+++ src/allegro_5/palfilt.c
@@ -17,7 +17,6 @@ void initcoef()
     ACoef2[1]=al_ftofix(0.58578643762667126000);
     ACoef2[2]=al_ftofix(0.29289321881333563000);
     BCoef2[2]=al_ftofix(0.17157287525380990000);
-    //rpclog("%08X %08X %08X %08X\n",ACoef2[0],ACoef2[1],ACoef2[2],BCoef2[2]);
 }
