+++ src/old_src/elkwin.rc
