+++ src/old_src/resources.h
